:bug: returns rejected Promise instead throws Error
It is a bad pattern to throw an error in a function whose return type is
Promise. When to handle multiple promises such as `Promise.all([...])`,
if there is a function that throws an error, unhandledrejection will
occur due to the rejection of other Promises.

In `Client.prototype.call()`, when session is closed, `this.#recv(...)`
returns a rejected Promise and `this.#session.send(...)` throws an Error,
then the rejected Promise is not handled anyone and unhundledrejection
is occured.

```
Unhandled rejection: Error: Session is not running
    at Session.recv (file:///work/vim/deno-vim-channel-command/session.ts:93:13)
    at Client.#recv (file:///work/vim/deno-vim-channel-command/client.ts:67:47)
    at Client.call (file:///work/vim/deno-vim-channel-command/client.ts:149:17)
    at Vim.call (file:///work/vim/denops.vim/denops/@denops-private/host/vim.ts:51:39)
    ...
```